440: Split voterGroups controller into separate files for each use case
