--- apps/backend/src/controllers/voterGroups.controllers.ts
@@ -1,127 +1,0 @@
-import {
-  response4XXObject,
-  type InsertableVoterGroup,
-  type Response400,
-  type Response403,
-  type Response404,
-  type Response500,
-  type SelectableUser,
-  type SelectableVoterGroup,
-} from '@repo/votura-validators';
-import type { Request, Response } from 'express';
-import jwt from 'jsonwebtoken';
-import { generateRSAKeyPair } from '../auth/generateJWTKeyPair.js';
-import { JWT_CONFIG } from '../auth/jwtConfig.js';
-import type { VoterJwtPayload } from '../auth/types.js';
-import { HttpStatusCode } from '../httpStatusCode.js';
-import {
-  createVoterGroup as createPersistentVoterGroup,
-  deleteVoterGroup as deletePersistentVoterGroup,
-  getVoterGroup,
-  getVoterGroupsForUser,
-  getVoterIdsForVoterGroup,
-  updateVoterGroup as updatePersistentVoterGroup,
-  updateVoterGroupPubKey,
-} from '../services/voterGroups.service.js';
-import { isBodyCheckValidationError } from './bodyChecks/bodyCheckValidationError.js';
-import {
-  validateInsertableVoterGroup,
-  validateUpdateableVoterGroup,
-} from './bodyChecks/voterGroupChecks.js';
-
-export const createVoterGroup = async (
-  req: Request,
-  res: Response<
-    SelectableVoterGroup | Response400 | Response403 | Response404 | Response500,
-    { user: SelectableUser }
-  >,
-): Promise<void> => {
-  const validationResult = await validateInsertableVoterGroup(req.body, res.locals.user.id);
-
-  if (isBodyCheckValidationError(validationResult)) {
-    res
-      .status(validationResult.status)
-      .json(response4XXObject.parse({ message: validationResult.message }));
-    return;
-  }
-
-  // If we reach this point, the request body is valid
-  const insertableVoterGroup: InsertableVoterGroup = validationResult;
-
-  // Proceed with creating the voter group
-  const voterGroup = await createPersistentVoterGroup(insertableVoterGroup, res.locals.user.id);
-  res.status(HttpStatusCode.created).send(voterGroup);
-};
-
-export const getVoterGroups = async (
-  _req: Request,
-  res: Response<SelectableVoterGroup[], { user: SelectableUser }>,
-): Promise<void> => {
-  const voterGroups = await getVoterGroupsForUser(res.locals.user.id);
-  res.status(HttpStatusCode.ok).json(voterGroups);
-};
-
-export const updateVoterGroup = async (
-  req: Request<{ voterGroupId: SelectableVoterGroup['id'] }>,
-  res: Response<
-    SelectableVoterGroup | Response400 | Response403 | Response404 | Response500,
-    { user: SelectableUser }
-  >,
-): Promise<void> => {
-  const validationResult = await validateUpdateableVoterGroup(req.body, res.locals.user.id);
-
-  if (isBodyCheckValidationError(validationResult)) {
-    res
-      .status(validationResult.status)
-      .json(response4XXObject.parse({ message: validationResult.message }));
-    return;
-  }
-
-  // If we reach this point, the request body is valid
-  const insertableVoterGroup: InsertableVoterGroup = validationResult;
-
-  // Proceed with updating the voter group
-  const voterGroup = await updatePersistentVoterGroup(
-    req.params.voterGroupId,
-    insertableVoterGroup,
-  );
-  res.status(HttpStatusCode.ok).send(voterGroup);
-};
-
-export const getSpecificVoterGroup = async (
-  req: Request<{ voterGroupId: SelectableVoterGroup['id'] }>,
-  res: Response<SelectableVoterGroup, { user: SelectableUser }>,
-): Promise<void> => {
-  const voterGroup = await getVoterGroup(req.params.voterGroupId);
-  res.status(HttpStatusCode.ok).json(voterGroup);
-};
-
-export const deleteVoterGroup = async (
-  req: Request<{ voterGroupId: SelectableVoterGroup['id'] }>,
-  res: Response<void, { user: SelectableUser }>,
-): Promise<void> => {
-  await deletePersistentVoterGroup(req.params.voterGroupId);
-  res.status(HttpStatusCode.noContent).send();
-};
-
-export const createVoterTokens = async (
-  req: Request<{ voterGroupId: SelectableVoterGroup['id'] }>,
-  res: Response<string[], { user: SelectableUser }>,
-): Promise<void> => {
-  const { privateKey, publicKey } = generateRSAKeyPair();
-
-  const voterIds = await getVoterIdsForVoterGroup(req.params.voterGroupId);
-  const voterTokens: string[] = [];
-  for (const voterId of voterIds) {
-    const tokenPayload: VoterJwtPayload = {
-      sub: voterId,
-    };
-    const votingToken = jwt.sign(tokenPayload, privateKey, {
-      algorithm: JWT_CONFIG.algorithm,
-    });
-    voterTokens.push(votingToken);
-  }
-  await updateVoterGroupPubKey(req.params.voterGroupId, publicKey);
-
-  res.status(HttpStatusCode.ok).send(voterTokens);
-};
